Add scoped component files for dashboard pages
