+++ app/javascript/components/stakeholder/StakeholderPage.jsx
@@ -5,57 +5,14 @@ import BusinessOutlinedIcon from '@mui/icons-material/BusinessOutlined';
 import HomeOutlinedIcon from '@mui/icons-material/HomeOutlined';
 import StorefrontOutlinedIcon from '@mui/icons-material/StorefrontOutlined';
 import Container from '@mui/material/Container';
-import Grid from '@mui/material/Grid';
 import { ThemeProvider, createTheme, responsiveFontSizes } from '@mui/material/styles';
-import Ch_Top20Vendors from '../Ch_Top20Vendors';
-import Ch_OrdersVolumeComparison from '../Ch_OrdersVolumeComparison';
 import Dashboard from '../Dashboard';
-import Panel from '../Panel';
+import Index from './Index';
+import Clients from './Clients';
+import Vendors from './Vendors';
 
 let theme = createTheme();
 theme = responsiveFontSizes(theme);
-
-const Index = (props) => {
-  return (
-    <Grid container spacing={3}>
-      <Grid item xs={12} md={12} lg={12}>
-        <Panel title="Example Panel">
-
-        </Panel>
-      </Grid>
-      <Grid item xs={12} md={12} lg={12}>
-        <Panel title='Top 20 Vendors'>
-          <Ch_Top20Vendors data={props.data}/>
-        </Panel>
-      </Grid>
-      <Grid item xs={12} md={12} lg={12}>
-        <Panel title='Orders volume comparison last month'>
-          <Ch_OrdersVolumeComparison data={props.data}/>
-        </Panel>
-      </Grid>
-    </Grid>
-  );
-}
-
-const Clients = (props) => {
-  return (
-    <Grid container spacing={3}>
-      <Grid item xs={12} md={12} lg={12}>
-        <Panel title="Clients"></Panel>
-      </Grid>
-    </Grid>
-  );
-}
-
-const Vendors = (props) => {
-  return (
-    <Grid container spacing={3}>
-      <Grid item xs={12} md={12} lg={12}>
-        <Panel title="Vendors"></Panel>
-      </Grid>
-    </Grid>
-  );
-}
 
 const StakeholderPage = (props) => {
   return (
